// Angular/DaycareSolutionSystem.MobileApp/src/app/api/generated/model/registeredActionDTO.ts
/**
 * My API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ActionDTO } from './actionDTO';
import { DayOfWeek } from './dayOfWeek';
import { PictureDTO } from './pictureDTO';


export interface RegisteredActionDTO { 
    id?: string;
    actionStartedDateTime?: Date;
    actionFinishedDateTime?: Date;
    plannedStartDateTime?: Date;
    day?: DayOfWeek;
    estimatedDurationMinutes?: number;
    action?: ActionDTO;
    clientActionSpecificDescription?: string;
    comment?: string;
    isCompleted?: boolean;
    isCanceled?: boolean;
    photo?: PictureDTO;
}